requirements.txt
pyaes==1.6.1
pyasn1==0.4.2
PySocks==1.6.8
rsa==3.4.2
Telethon==0.18.0.3
